refactor: reorganize donation components and update import paths
- Moved donation-related components into a dedicated `features/donation` directory for improved organization and maintainability.
- Updated import paths across multiple files to reflect this new structure, enhancing clarity and consistency in the codebase.
