+++ src/app/(app)/(stacked-layout)/donate/page.tsx
@@ -3,10 +3,10 @@ import { Handshake, Trophy } from "lucide-react";
 
 import { siteConfig } from "@/config/site";
 import { Separator } from "@/components/ui/separator";
-import { DonateBadgeShowcase } from "@/components/donate/donate-badge-showcase";
-import { DonateButton } from "@/components/donate/donate-button";
-import { DonateEmailReminder } from "@/components/donate/donate-email-reminder";
-import { DonateFeature } from "@/components/donate/donate-feature";
+import { DonateBadgeShowcase } from "@/components/features/donation/donate-badge-showcase";
+import { DonateButton } from "@/components/features/donation/donate-button";
+import { DonateEmailReminder } from "@/components/features/donation/donate-email-reminder";
+import { DonateFeature } from "@/components/features/donation/donate-feature";
 import { ButtonWithIcon } from "@/components/shared/button-with-icon";
 import { Icons } from "@/components/shared/icons";
 import { Typography } from "@/components/shared/typography";
